msm: camera: ife: implicit RDI allocation
Each of the IFE multiple RDIs are equivalent but current userspace
interface allows only explicitly requested (by id) RDI to be
allocated. This causes allocation of multiple IFEs, while one can do
the job, and can reduce the functionality of IFE driver in multicamera
usecases. With this change, a check is made if an IFE has another
suitable free interface, if requested is already busy and allocates it
instead. Also communicates with userspace the id of allocated
interface.

CRs-Fixed: 3081728
Change-Id: Icd118a651d15544d64911d8705deaec5c51711a9
Signed-off-by: Petar Sivenov <quic_c_psiven@quicinc.com>